requirements.txt
gradio
yt-dlp
faster-whisper
ctranslate2==4.4.0
torch==2.5
torchaudio==2.5
google-genai
psycopg2-binary
google-generativeai
langchain
scipy
langchain_google_genai
moviepy
opencv-python
soundfile
pydub
langgraph
ipython
transformers